+++ CHANGELOG.md
@@ -5,8 +5,6 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
-## [0.21.0] - 2025-04-11
-
 ### Changed
 - **BREAKING**: Updated endpoints in `pw.xpacks.llm.question_answering.BaseRAGQuestionAnswerer`:
   - Deprecated: `/v1/pw_list_documents`, `/v1/pw_ai_answer`
@@ -15,7 +13,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
   - `pw_ai_summary` -> `summarize`
   - `pw_ai_answer` -> `answer`
   - `pw_list_documents` -> `list_documents`
-  
+- When `pw.io.deltalake.write` creates a table, it also stores its metadata in the columns of the created Delta table. This metadata can be used by Pathway when reading the table with `pw.io.deltalake.read` if no `schema` is specified.
+- The `schema` parameter is now optional for `pw.io.deltalake.read`. If the table was created by Pathway and the `schema` was not specified by user, it is read from the table metadata.
 
 ## [0.20.1] - 2025-03-07